# packages/shared-types/src/error-codes.ts
export const ERROR_CODES = {
  AUTH_UNAUTHORIZED: 'AUTH_UNAUTHORIZED',
  AUTH_DEVICE_UNTRUSTED: 'AUTH_DEVICE_UNTRUSTED',
  PERM_ROLE_FORBIDDEN: 'PERM_ROLE_FORBIDDEN',
  PERM_NO_USE_PERMISSION: 'PERM_NO_USE_PERMISSION',
  INSTALL_CONFLICT: 'INSTALL_CONFLICT',
  INSTALL_RECORD_STATUS_CONFLICT: 'INSTALL_RECORD_STATUS_CONFLICT',
  TICKET_NOT_FOUND: 'TICKET_NOT_FOUND',
  TICKET_EXPIRED: 'TICKET_EXPIRED',
  TICKET_ALREADY_CONSUMED: 'TICKET_ALREADY_CONSUMED',
  SEARCH_SERVICE_DEGRADED: 'SEARCH_SERVICE_DEGRADED',
  SEARCH_QUERY_INVALID: 'SEARCH_QUERY_INVALID',
  INDEX_BUILD_FAILED: 'INDEX_BUILD_FAILED'
} as const;

export type ErrorCode = (typeof ERROR_CODES)[keyof typeof ERROR_CODES];

export interface ErrorResponse {
  code: ErrorCode | string;
  message: string;
  requestId: string;
  traceId: string;
}